refactor: 예외 처리 책임 이동 및 누락된 예외 처리 추가

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -3,17 +3,22 @@ package roomescape.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.dto.ReservationTimeRequestDto;
 import roomescape.dto.ReservationTimeResponseDto;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 
 @Service
 public class ReservationTimeService {
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ReservationTimeService(final ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(final ReservationTimeRepository reservationTimeRepository,
+                                  final ReservationRepository reservationRepository) {
         this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     @Transactional
@@ -32,6 +37,10 @@ public class ReservationTimeService {
 
     @Transactional
     public void delete(final long id) {
+        final List<Reservation> reservations = reservationRepository.findByTimeId(id);
+        if (!reservations.isEmpty()) {
+            throw new IllegalArgumentException("Reservation exists with this time");
+        }
         reservationTimeRepository.deleteById(id);
     }
 }
